Add fallback to `resolveByTwigPath` to handle missing templates (#27)
### Overview

This PR introduces a new utility function `resolveTemplate`. The
function is designed to search for a template file and return the first
match if one exists.

### Functionality

The resolveTemplate function works as follows:

1. It first checks if the provided path points to a file. If it does, it
returns the path.
2. If the path does not point to a file, it appends each of the
predefined extensions to the path and checks if a file exists at the
resulting path. If it finds a file, it returns the new path.
3. If no file is found with the previous steps, it appends each
combination of predefined index filenames and extensions to the path and
checks if a file exists at the resulting path. If it finds a file, it
returns the new path.
4. If no file is found after all checks, it returns null.

closes #26